+++ minecraft/src/net/minecraft/server/level/ServerLevel.java
@@ -1153,7 +1153,7 @@ public class ServerLevel extends Level {
 
 	public <T extends ParticleOptions> int sendParticles(T particleOptions, double d, double e, double f, int i, double g, double h, double j, double k) {
 		ClientboundLevelParticlesPacket clientboundLevelParticlesPacket = new ClientboundLevelParticlesPacket(
-			particleOptions, false, (float)d, (float)e, (float)f, (float)g, (float)h, (float)j, (float)k, i
+			particleOptions, false, d, e, f, (float)g, (float)h, (float)j, (float)k, i
 		);
 		int l = 0;
 
@@ -1170,7 +1170,7 @@ public class ServerLevel extends Level {
 	public <T extends ParticleOptions> boolean sendParticles(
 		ServerPlayer serverPlayer, T particleOptions, boolean bl, double d, double e, double f, int i, double g, double h, double j, double k
 	) {
-		Packet<?> packet = new ClientboundLevelParticlesPacket(particleOptions, bl, (float)d, (float)e, (float)f, (float)g, (float)h, (float)j, (float)k, i);
+		Packet<?> packet = new ClientboundLevelParticlesPacket(particleOptions, bl, d, e, f, (float)g, (float)h, (float)j, (float)k, i);
 		return this.sendParticles(serverPlayer, bl, d, e, f, packet);
 	}
