+++ minecraft/src/net/minecraft/world/level/block/MyceliumBlock.java
@@ -20,9 +20,9 @@ public class MyceliumBlock extends SpreadingSnowyDirtBlock {
 		if (random.nextInt(10) == 0) {
 			level.addParticle(
 				ParticleTypes.MYCELIUM,
-				(double)((float)blockPos.getX() + random.nextFloat()),
-				(double)((float)blockPos.getY() + 1.1F),
-				(double)((float)blockPos.getZ() + random.nextFloat()),
+				(double)blockPos.getX() + (double)random.nextFloat(),
+				(double)blockPos.getY() + 1.1,
+				(double)blockPos.getZ() + (double)random.nextFloat(),
 				0.0,
 				0.0,
 				0.0
