+++ minecraft/src/net/minecraft/world/level/block/NetherPortalBlock.java
@@ -124,9 +124,9 @@ public class NetherPortalBlock extends Block {
 		}
 
 		for (int i = 0; i < 4; i++) {
-			double d = (double)((float)blockPos.getX() + random.nextFloat());
-			double e = (double)((float)blockPos.getY() + random.nextFloat());
-			double f = (double)((float)blockPos.getZ() + random.nextFloat());
+			double d = (double)blockPos.getX() + (double)random.nextFloat();
+			double e = (double)blockPos.getY() + (double)random.nextFloat();
+			double f = (double)blockPos.getZ() + (double)random.nextFloat();
 			double g = ((double)random.nextFloat() - 0.5) * 0.5;
 			double h = ((double)random.nextFloat() - 0.5) * 0.5;
 			double j = ((double)random.nextFloat() - 0.5) * 0.5;
